requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.1
django-phone-auth==0.3.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
flake8==6.1.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
phonenumbers==8.13.18
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.1.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
